[Refactor] Define TabletsChannel as an abstract class
